False positive 2 (#63)
Fixes #59
* separated context variables into more files
* added system_variable
* added system variables into scope stack
* *PROCESS
* bad highlighting in macro params
* adjust_range bug
* base implementation of AREAD - just defined variable symbol
* removed false expression check
* lookahead syntax error bug
* highlighting bug in macro label
* package indent
* added missing includes

+++ parser_library/src/expressions/character_expression.cpp
@@ -49,7 +49,7 @@ void character_expression::append(std::string v)
 
 int ebcdic_compare(const std::string& lhs, const std::string& rhs)
 {
-	int diff = lhs.size() - rhs.size();
+	int diff = (int)lhs.size() - (int)rhs.size();
 
 	if (diff != 0)
 		return diff;
